Remove BitcoindV21RpcClientTest and move test cases to BitcoindV24RpcClientTest
